LayoutTests/fast/css/color-valid-contrast-color-function.html

<!DOCTYPE html>
<html>
<head>
<title>Test CSS Color Module 5 contrast-color() parsing</title>
<meta charset="utf-8">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../imported/w3c/web-platform-tests/css/support/parsing-testcommon.js"></script>
<style>
  #container {
    container-type: inline-size;
    width: 10px;
    color: pink;
  }
</style>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<script>

// Test basic legacy colors
test_valid_value("background-color", "contrast-color(white)");
test_valid_value("background-color", "contrast-color(white max)");
test_valid_value("background-color", "contrast-color(black)");
test_valid_value("background-color", "contrast-color(black max)");
test_valid_value("background-color", "contrast-color(pink)");
test_valid_value("background-color", "contrast-color(pink max)");

// Test non-legacy colors
test_valid_value("background-color", "contrast-color(color(srgb 1 0 1 / 0.5))");
test_valid_value("background-color", "contrast-color(color(srgb 1 0 1 / 0.5) max)");
test_valid_value("background-color", "contrast-color(lab(0.2 0.5 0.2))");
test_valid_value("background-color", "contrast-color(lab(0.2 0.5 0.2) max)");

// Test out-of-gamut colors
test_valid_value("background-color", "contrast-color(color(srgb 10 10 10))");
test_valid_value("background-color", "contrast-color(color(srgb 10 10 10) max)");
test_valid_value("background-color", "contrast-color(color(srgb -10 -10 -10))");
test_valid_value("background-color", "contrast-color(color(srgb -10 -10 -10) max)");

// Test nested contrast-color
test_valid_value("background-color", "contrast-color(contrast-color(pink))");
test_valid_value("background-color", "contrast-color(contrast-color(pink max))");
test_valid_value("background-color", "contrast-color(contrast-color(pink) max)");
test_valid_value("background-color", "contrast-color(contrast-color(pink max) max)");

// Test currentColor
test_valid_value("background-color", "contrast-color(currentcolor)");
test_valid_value("background-color", "contrast-color(currentcolor max)");

// Test color using calc().
test_valid_value("background-color", "contrast-color(color(srgb calc(0.5) calc(1 + (sign(20cqw - 10px) * 0.5)) 1 / .5))", "contrast-color(color(srgb calc(0.5) calc(1 + (0.5 * sign(20cqw - 10px))) 1 / 0.5))");
test_valid_value("background-color", "contrast-color(color(srgb calc(0.5) calc(1 + (sign(20cqw - 10px) * 0.5)) 1 / .5) max)", "contrast-color(color(srgb calc(0.5) calc(1 + (0.5 * sign(20cqw - 10px))) 1 / 0.5) max)");

</script>
</body>
</html>
